+++ conf/e4/config.json
@@ -2,7 +2,7 @@
 	"include": [
 		"keywords.json",
         "prefixes.json",
-        "e3/terrains.xml"
+        "e3/terrains.json"
 	],
     "settings": {
         "game.id": 4,
