+++ s/runtests
@@ -19,6 +19,7 @@ cd $ROOT
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests.lua
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e2.lua
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e3.lua
+$ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e4.lua
 rm -rf data reports orders.txt
 
 cd $OLDWPD
