+++ scripts/run-tests-e2.lua
@@ -15,6 +15,7 @@ require 'eressea'
 require 'eressea.xmlconf'
 require 'eressea.path'
 require 'tests.e2'
+require 'tests.xmas'
 require 'lunit'
 
 rules = require('eressea.' .. config.rules)
