+++ tests/runtests.bat
@@ -5,5 +5,6 @@ SET SERVER=%BUILD%\eressea.exe
 %SERVER% ..\scripts\run-tests.lua
 %SERVER% ..\scripts\run-tests-e2.lua
 %SERVER% ..\scripts\run-tests-e3.lua
+%SERVER% ..\scripts\run-tests-e4.lua
 PAUSE
 RMDIR /s /q reports
